Refactor: changed file structure of config
