+++ src/main/kotlin/com/mmcm/projectocp/backend/spring/config/SecurityConfig.kt
@@ -1,5 +1,7 @@
 package com.mmcm.projectocp.backend.spring.config
 
+import com.mmcm.projectocp.backend.spring.config.filter.JwtAuthenticationFilter
+import com.mmcm.projectocp.backend.spring.config.handler.CustomLogoutSuccessHandler
 import com.mmcm.projectocp.backend.spring.config.service.*
 import com.mmcm.projectocp.backend.spring.domain.repository.RefreshTokenRepository
 import com.mmcm.projectocp.backend.spring.domain.repository.UserRepository
